requirements.txt
customtkinter == 5.2.2
openpyxl == 3.1.5
pandas == 2.2.3
pydantic-settings == 2.7.1
pymongo == 4.11
xlsxwriter == 3.2.2
python-dotenv==1.0.1
pydantic==2.10.6
dnspython==2.7.0
pip==24.3.1
filelock==3.17.0
pytz==2025.1
numpy==2.2.2
python-dateutil==2.9.0.post0
six==1.17.0
et_xmlfile==2.0.0
typing_extensions==4.12.2
pydantic_core==2.27.2
annotated-types==0.7.0
packaging==24.2
darkdetect==0.8.0
PyYAML==6.0.2
